fusion: double accumulators for heave and sea state
The audit ranked this as "measure first, act on the result" because a
worst-case ULP bound had over-predicted the heave error by ~19x once
already.  Measured, the bound was right this time, and worse than the
finding that prompted the audit.

BENCH_SWEEP_PRECISION runs each estimator against a double-precision
reference of the same arithmetic, driven with the SAME float-rounded
inputs so input quantisation is common-mode and only accumulator
precision is left.  The axis is dt/tau -- the EMA gain -- because that is
what approaches float32 epsilon, and it shrinks with BOTH the sample rate
and the window.

Sea state, float-vs-double before:

    dt/tau                      Hs error   Tz error
    1.0e-5   (833 Hz/120 s)      0.000 %    0.000 %   <- shipped default
    2.1e-7   (8 kHz/600 s)       0.128 %    0.491 %
    5.2e-8   (32 kHz/600 s)      1.823 %    0.982 %
    2.6e-8   (32 kHz/1200 s)    17.995 %    6.865 %

That last row is a valid-flagged 18 % under-report of significant wave
height, and 10-20 min records are what fusion.c's own comment calls
oceanographic practice.

Heave was worse, and only showed up because the sweep was given its own
tau axis.  heave_tau_s enters as leak = dt/tau on both integrators and as
alpha = tau/(tau+dt) on the output high-pass.  When the leak goes under
half an ULP it becomes an exact no-op and the LEAKY integrator turns into
a plain one -- unbounded:

    32 kHz / tau  12 s     0.225 %      <- shipped default
    32 kHz / tau  60 s     1.225 %
    32 kHz / tau 300 s     7.263 %
    32 kHz / tau 900 s  6543974 %       <- 253 m of "heave", flagged valid

Both are now double: seastate_t's six mean/variance pairs and its alpha,
heave_t's vel/disp/disp_prev/hp_y with leak, alpha and the gravity
subtraction in double.  The gravity removal is a cancellation of ~9.8
against ~9.8 down to a signal three orders smaller, so it belongs on the
wide side; the rotation stays float because that is the accel sample's own
precision and widening it would flatter the result without changing the
hardware.  Every cell in both tables is 0.000 % after.

Wire types and accessors are untouched -- heave_update and the six
seastate getters still return float.

The default-on sweeps confirm it independently.  test_seastate_across_rates
was flat at 1.015 % Hs from 100 Hz to 16 kHz and then ticked up to 1.046 %
at 32 kHz; that endpoint was recorded yesterday as float32 accumulation and
it is now 1.016 %, i.e. the plateau continues to 32 kHz with the uptick
gone.  Worst period 0.161 % -> 0.125 %.  Heave's worst is unchanged at
1.157 % because it sits at 12 Hz, where the limit is resolution, not
precision.

Two guards.  test_leaks_survive_a_tiny_dt_over_tau pins the arithmetic
rather than the symptom -- reproducing the blow-up needs 25 filter
constants at 32 kHz, 720 M iterations, while the mechanism is visible in
one update.  Reverting both types fails 4 of its 5 assertions.  Note two
of those assertions had to be quantitative: float does not round the leak
to zero, it rounds it to a whole ULP, so "it still decayed" passes while
the decay RATE is 72 % wrong.

No startup warning, contrary to the plan.  Once both accumulators are
double the cliff moves to dt/tau ~ 1e-16, which no rate and window can
reach, so a thin-gain warning could never fire -- and a check that cannot
fire is worse than none.

Also documents why imu.c's bias-init variance may keep the
cancellation-prone E[x^2]-mean^2 form: double, bounded window, values
~1e-3 rad/s, so three digits lost out of sixteen.  In float it would keep
about four and the 0.5 deg/s motion threshold would be reading noise.

5516 assertions -> 5522.

+++ include/fusion.h
@@ -274,10 +274,31 @@ float mekf_ema_alpha(float dt_s, float tau_s);
 typedef struct {
     float tau;       /* leak / high-pass time constant, s */
     float dt;        /* sample period, s */
-    float vel;       /* leaked vertical velocity, m/s (down +) */
-    float disp;      /* leaked vertical displacement, m (down +) */
-    float disp_prev; /* previous displacement (high-pass state) */
-    float hp_y;      /* high-passed displacement (down +) */
+    /*
+     * Double, for a sharper version of the reason seastate_t gives.  Both
+     * stages are leaky — vel -= vel*(dt/tau) — and the leak is what bounds
+     * the drift of a double integration of accelerometer noise.  In float32
+     * that leak becomes a no-op once dt/tau falls below half an ULP, and the
+     * filter silently turns into a PLAIN double integrator.  The output
+     * high-pass fails the same way: alpha = tau/(tau+dt) rounds to exactly
+     * 1.0 and the exact zero at DC is gone.
+     *
+     * Measured against a double reference of the same arithmetic
+     * (BENCH_SWEEP_PRECISION), recovered amplitude error:
+     *
+     *     32 kHz / tau  12 s   0.225 %      <- shipped default
+     *     32 kHz / tau  60 s   1.225 %
+     *     32 kHz / tau 300 s   7.263 %
+     *     32 kHz / tau 900 s   6543974 %    <- 253 m of "heave"
+     *
+     * That last row is not a degraded reading, it is an unbounded integrator
+     * with FLAG_HEAVE_VALID set.  Double moves the cliff to dt/tau ~ 1e-16,
+     * which nothing reaches.  heave_update still returns float.
+     */
+    double vel;       /* leaked vertical velocity, m/s (down +) */
+    double disp;      /* leaked vertical displacement, m (down +) */
+    double disp_prev; /* previous displacement (high-pass state) */
+    double hp_y;      /* high-passed displacement (down +) */
     /*
      * Settling is counted in SAMPLES, not summed in seconds.  A float
      * accumulating a constant dt stops advancing once dt drops below half an
@@ -314,12 +335,32 @@ float heave_update(heave_t *h, const float q[4], const float accel[3]);
 typedef struct {
     float tau;       /* averaging time constant, s */
     float dt;        /* sample period, s */
-    float h_mean,  h_var;    /* heave, m / m² */
-    float hr_mean, hr_var;   /* heave rate, m/s / (m/s)² */
-    float r_mean,  r_var;    /* roll, rad / rad² (mean = steady heel) */
-    float rr_mean, rr_var;   /* roll rate, rad/s / (rad/s)² */
-    float p_mean,  p_var;    /* pitch, rad / rad² (mean = steady trim) */
-    float pr_mean, pr_var;   /* pitch rate, rad/s / (rad/s)² */
+    /*
+     * The accumulators are DOUBLE, and this is the one place in the filter
+     * where that is load-bearing rather than tidy.  The EMA gain is dt/tau,
+     * so it shrinks with the sample rate AND with the window; at 32 kHz with
+     * a 1200 s window it is 2.6e-8, well under float32's 1.19e-7 epsilon, and
+     * `mean += alpha*d` stops making progress.  Measured against a
+     * double-precision reference of the same arithmetic (BENCH_SWEEP_PRECISION):
+     *
+     *     dt/tau        Hs error   Tz error
+     *     1.0e-5  (833 Hz/120 s)      0.000 %    0.000 %   <- shipped default
+     *     2.1e-7  (8 kHz/600 s)       0.128 %    0.491 %
+     *     5.2e-8  (32 kHz/600 s)      1.823 %    0.982 %
+     *     2.6e-8  (32 kHz/1200 s)    17.995 %    6.865 %
+     *
+     * — and that last row is a valid-flagged 18 % under-report of significant
+     * wave height, not a missing reading.  Both long windows are documented
+     * practice (10–20 min records), so the combination is reachable.  Double
+     * moves the same cliff out past dt/tau ~ 1e-16, which no rate reaches.
+     * The accessors still return float; only the accumulation is widened.
+     */
+    double h_mean,  h_var;   /* heave, m / m² */
+    double hr_mean, hr_var;  /* heave rate, m/s / (m/s)² */
+    double r_mean,  r_var;   /* roll, rad / rad² (mean = steady heel) */
+    double rr_mean, rr_var;  /* roll rate, rad/s / (rad/s)² */
+    double p_mean,  p_var;   /* pitch, rad / rad² (mean = steady trim) */
+    double pr_mean, pr_var;  /* pitch rate, rad/s / (rad/s)² */
     /* Samples, not summed seconds — see heave_t for why.  This one is the
      * reachable case: `settled` needs 2·wave_tau_s, and wave_tau_s has no
      * upper bound while the window this estimator is built for is minutes. */
